change logic work whith point type
