+++ src/constants.js
@@ -1,0 +1,24 @@
+const insertPosition = {
+  BEFORE_BEGIN: `beforebegin`,
+  AFTER_BEGIN: `afterbegin`,
+  BEFORE_END: `beforeend`,
+  AFTER_END: `afterend`
+};
+
+const pointTypeToPretext = {
+  'taxi': `Taxi to`,
+  'bus': `Bus to`,
+  'train': `Train to`,
+  'ship': `Ship to`,
+  'transport': `Transport to`,
+  'drive': `Drive to`,
+  'flight': `Flight to`,
+  'check-in': `Check in`,
+  'sightseeing': `Sightseeing in`,
+  'restaurant': `Restaurant in`,
+};
+
+export {
+  insertPosition,
+  pointTypeToPretext
+};
